Move orhelper logger out of file scope
Since the orhelper logger is created in the file scope of `_orhelper.py`, it is created when orhelper is imported. This means that to assign settings to that logger, clients have to import `logging` themselves, get the logger that orhelper will be assigned with `logging.getLogger('orhelper._orhelper')`, assign settings, and then finally import orhelper. This is quite ugly in client code and it would be much nicer if the logger was created after import, i.e. in the `OpenRocketInstance` constructor.